Fix builds on PAX hardenned systems
PAX flags on RWX memory protections for elf2bin's machinations,
resulting in build failures unless one manually sets the FS xattr
or PAX header flags on the binary.

This is the low road - check for presence of paxctl binary in the
PATH, use it to convert the ELF headers, and set "em" flags which
allow unsafe memory operations to be performed.
The high road would be to check sysctls and set FS XATTRs, but it
would add unecessary complexity as well as privilege requirements
generally undesirable for build enivironments (root builds == bad).

TPE and other restrictions commonly present in grsec environments
are not addressed here.