Parallel and dynamic rules for building Mettle.
Based on [PR #58](https://github.com/rapid7/mettle/pull/58), this is just a different take on fixing #43.  When `make` is invoked, the current list of target architectures is read from the ARCHES file and rules are created for doing parallel build/clean/distclean operations on them.

NOTE: I tested the make rules look and act correctly, but did not try an actual build.