Make background and debug options take a non-optional parameter
Today we learned that optional parameters as parsed by getopt_long are rather lame to actually handle from a user's point-of-view. This is because they require you to understand that in order to pass the optional argument, you have to purposely avoid including a space between the option flag and the value, in contrast to non-optional parameters, which _do_ allow a space.

This allows for all of these variants to work as you would actually expect:

```
./mettle -d 1
./mettle -d1
./mettle -b 1
./mettle -b0
```