Use correct format string hinting on MinGW
When compiling for Windows targets, the `printf` format string archetype
will warn about formats not available natively on Windows. We were
already specifying __USE_MINGW_ANSI_STDIO to invoke the compatibility
layer in MinGW, but we still needed to update third-party code to use
the correct hint, `gnu_printf` as set in the MinGW macro.

https://sourceforge.net/p/mingw-w64/wiki2/gnu%20printf/#user-printf-and-scanf