Fix magic numbers behind first packet detection
Mettle was failing to function because the packet detection code relies
on a certain amount of data. Given that the packet sizes decreased, the
size value had to decrease too. I wasn't aware of this at the time and
it took a bit to track down.

I've adjusted this code a little so that we don't have magic numbers in
there any more. The labels on the definitions should at least give
future developers an idea of what those numbers mean.

I'd honestly prefer to not do it this way, but this will have to do for
now. It's important to note that when the PR for changing PEM to DER
gets landed, this will break again as the packet will be smaller.