set AT_SYSINFO_EHDR so that vDSO is mapped in
The vDSO (virtual dynamic shared object) is a shared library that is
mapped in on Linux kernels for performance reasons. Instead of doing
context-switches for f.e. gettimeofday() one can stay in userland and
the kernel will take care of updating the values in the memory mapped
region in userland.

This patch will copy the pointer to that region properly over to the new
auxiliary vector such that the resulting reflected binary has the
ability to perform vDSO calls. For example a simple "hello world" binary
written in Go will already need this else it will crash with a SIGSEGV
on x86-64 Linux systems.

To test we create a simple hello world in Golang.

$ cat > test.go << EOF
package main
import "fmt"
func main() {
fmt.Println("hello world")
}
EOF
$ go build test.go && ./test
hello world

Before adding the patch we get the following:

$ ulimit -c unlimited
$ ./build/linux.x86_64/libreflect/examples/noexec ./test
Segmentation fault (core dumped)
$ gdb ./build/linux.x86_64/bin/noexec core
...
(gdb) x/i $rip
=> 0xffffffffff600000:  Cannot access memory at address
(gdb)

After adding the patch we get this:

$ ./build/linux.x86_64/libreflect/examples/noexec ./test
hello world